Fix connection string and tickets\products js files (all renamed to tickets)
