+++ src/config/config.js
@@ -1,7 +1,8 @@
 require("dotenv").config();
+
 const config = {
   PORT: process.env.PORT || 5000,
-  DB_CONNECTION: `mongodb+srv://nivnet:NivNiv996@cluster0.j2yahvt.mongodb.net/anyticket?retryWrites=true&w=majority`,
+  DB_CONNECTION: `mongodb+srv://${process.env.USER_NAME}:${process.env.USER_PASSWORD}@swappi.1g75bvc.mongodb.net/Swappi`,
   SECRET: "zsocstmp",
 
   SALT: 10,
